add observation deserialization
